+++ crates/miden-testing/src/kernel_tests/tx/test_note.rs
@@ -6,7 +6,7 @@ use alloc::vec::Vec;
 use anyhow::Context;
 use miden_lib::account::wallets::BasicWallet;
 use miden_lib::errors::MasmError;
-use miden_lib::errors::tx_kernel_errors::ERR_NOTE_ATTEMPT_TO_ACCESS_NOTE_SENDER_FROM_INCORRECT_CONTEXT;
+use miden_lib::errors::tx_kernel_errors::ERR_NOTE_ATTEMPT_TO_ACCESS_NOTE_METADATA_WHILE_NO_NOTE_BEING_PROCESSED;
 use miden_lib::testing::mock_account::MockAccountExt;
 use miden_lib::testing::note::NoteBuilder;
 use miden_lib::transaction::TransactionKernel;
@@ -48,33 +48,47 @@ use crate::{
     TransactionContext,
     TransactionContextBuilder,
     TxContextInput,
-    assert_execution_error,
     assert_transaction_executor_error,
 };
 
 #[test]
-fn test_get_sender_no_sender() -> anyhow::Result<()> {
-    let tx_context = TransactionContextBuilder::with_existing_mock_account().build()?;
+fn test_get_sender_fails_from_tx_script() -> anyhow::Result<()> {
+    // Creates a mockchain with an account and a note
+    let mut builder = MockChain::builder();
+    let account = builder.add_existing_wallet(Auth::BasicAuth)?;
+    let p2id_note = builder.add_p2id_note(
+        ACCOUNT_ID_SENDER.try_into().unwrap(),
+        account.id(),
+        &[FungibleAsset::mock(150)],
+        NoteType::Public,
+    )?;
+    let mut mock_chain = builder.build()?;
+    mock_chain.prove_next_block()?;
+
     // calling get_sender should return sender
     let code = "
-        use.$kernel::memory
-        use.$kernel::prologue
         use.miden::note
 
         begin
-            exec.prologue::prepare_transaction
-
-            # force the current input note pointer to 0
-            push.0 exec.memory::set_current_input_note_ptr
-
-            # get the sender
+            # try to get the sender from transaction script
             exec.note::get_sender
         end
         ";
+    let tx_script = ScriptBuilder::default()
+        .compile_tx_script(code)
+        .context("failed to compile tx script")?;
 
-    let process = tx_context.execute_code(code);
+    let tx_context = mock_chain
+        .build_tx_context(TxContextInput::AccountId(account.id()), &[p2id_note.id()], &[])?
+        .tx_script(tx_script)
+        .build()?;
 
-    assert_execution_error!(process, ERR_NOTE_ATTEMPT_TO_ACCESS_NOTE_SENDER_FROM_INCORRECT_CONTEXT);
+    let result = tx_context.execute_blocking();
+    assert_transaction_executor_error!(
+        result,
+        ERR_NOTE_ATTEMPT_TO_ACCESS_NOTE_METADATA_WHILE_NO_NOTE_BEING_PROCESSED
+    );
+
     Ok(())
 }
 
@@ -112,79 +126,6 @@ fn test_get_sender() -> anyhow::Result<()> {
     let sender = tx_context.input_notes().get_note(0).note().metadata().sender();
     assert_eq!(process.stack.get(0), sender.prefix().as_felt());
     assert_eq!(process.stack.get(1), sender.suffix());
-    Ok(())
-}
-
-#[test]
-fn test_get_vault_data() -> anyhow::Result<()> {
-    let tx_context = {
-        let mut builder = MockChain::builder();
-        let account = builder.add_existing_wallet(crate::Auth::BasicAuth)?;
-        let p2id_note_1 = builder.add_p2id_note(
-            ACCOUNT_ID_SENDER.try_into().unwrap(),
-            account.id(),
-            &[FungibleAsset::mock(150)],
-            NoteType::Public,
-        )?;
-        let p2id_note_2 = builder.add_p2id_note(
-            ACCOUNT_ID_SENDER.try_into().unwrap(),
-            account.id(),
-            &[FungibleAsset::mock(300)],
-            NoteType::Public,
-        )?;
-        let mut mock_chain = builder.build()?;
-        mock_chain.prove_next_block()?;
-
-        mock_chain
-            .build_tx_context(
-                TxContextInput::AccountId(account.id()),
-                &[],
-                &[p2id_note_1, p2id_note_2],
-            )?
-            .build()?
-    };
-
-    let notes = tx_context.input_notes();
-
-    // calling get_assets_info should return assets info
-    let code = format!(
-        "
-        use.std::sys
-
-        use.$kernel::prologue
-        use.$kernel::note
-
-        begin
-            exec.prologue::prepare_transaction
-
-            # get the assets info about note 0
-            exec.note::get_assets_info
-
-            # assert the assets data is correct
-            push.{note_0_asset_commitment} assert_eqw
-            push.{note_0_num_assets} assert_eq
-
-            # increment current input note pointer
-            exec.note::increment_current_input_note_ptr
-
-            # get the assets info about note 1
-            exec.note::get_assets_info
-
-            # assert the assets data is correct
-            push.{note_1_asset_commitment} assert_eqw
-            push.{note_1_num_assets} assert_eq
-
-            # truncate the stack
-            exec.sys::truncate_stack
-        end
-        ",
-        note_0_asset_commitment = notes.get_note(0).note().assets().commitment(),
-        note_0_num_assets = notes.get_note(0).note().assets().num_assets(),
-        note_1_asset_commitment = notes.get_note(1).note().assets().commitment(),
-        note_1_num_assets = notes.get_note(1).note().assets().num_assets(),
-    );
-
-    tx_context.execute_code(&code)?;
     Ok(())
 }
 
@@ -345,18 +286,23 @@ fn test_get_inputs() -> anyhow::Result<()> {
             .build()?
     };
 
-    fn construct_input_assertions(note: &Note) -> String {
+    fn construct_inputs_assertions(note: &Note) -> String {
         let mut code = String::new();
-        for input_chunk in note.inputs().values().chunks(WORD_SIZE) {
-            let mut input_word = EMPTY_WORD;
-            input_word.as_mut_slice()[..input_chunk.len()].copy_from_slice(input_chunk);
+        for inputs_chunk in note.inputs().values().chunks(WORD_SIZE) {
+            let mut inputs_word = EMPTY_WORD;
+            inputs_word.as_mut_slice()[..inputs_chunk.len()].copy_from_slice(inputs_chunk);
 
             code += &format!(
-                "
-                # assert the input is correct
-                dup padw movup.4 mem_loadw push.{input_word} assert_eqw push.4 add
-                ",
-                input_word = input_word
+                r#"
+                # assert the inputs are correct
+                # => [dest_ptr]
+                dup padw movup.4 mem_loadw push.{inputs_word} assert_eqw.err="inputs are incorrect"
+                # => [dest_ptr]
+                
+                push.4 add
+                # => [dest_ptr+4]
+                "#,
+                inputs_word = inputs_word
             );
         }
         code
@@ -378,7 +324,7 @@ fn test_get_inputs() -> anyhow::Result<()> {
             exec.note_internal::prepare_note
             # => [note_script_root_ptr, NOTE_ARGS, pad(11)]
 
-            # drop the note inputs
+            # clean the stack
             dropw dropw dropw dropw
             # => []
 
@@ -391,17 +337,17 @@ fn test_get_inputs() -> anyhow::Result<()> {
             dup eq.{NOTE_0_PTR} assert
             # => [dest_ptr]
 
-            # apply note 1 input assertions
-            {input_assertions}
+            # apply note 1 inputs assertions
+            {inputs_assertions}
             # => [dest_ptr]
 
-            # clean the pointer
+            # clear the stack
             drop
             # => []
         end
         ",
         num_inputs = note0.inputs().num_values(),
-        input_assertions = construct_input_assertions(note0),
+        inputs_assertions = construct_inputs_assertions(note0),
         NOTE_0_PTR = 100000000,
     );
 
@@ -409,12 +355,12 @@ fn test_get_inputs() -> anyhow::Result<()> {
     Ok(())
 }
 
-/// This test checks the scenario when an input note has exactly 8 input values, and the transaction
+/// This test checks the scenario when an input note has exactly 8 inputs, and the transaction
 /// script attempts to load the inputs to memory using the `miden::note::get_inputs` procedure.
 ///
-/// Previously this setup was leading to the incorrect number of note input values computed during
-/// the `get_inputs` procedure, see the
-/// [issue #1363](https://github.com/0xMiden/miden-base/issues/1363) for more details.
+/// Previously this setup was leading to the incorrect number of note inputs computed during the
+/// `get_inputs` procedure, see the [issue #1363](https://github.com/0xMiden/miden-base/issues/1363)
+/// for more details.
 #[test]
 fn test_get_exactly_8_inputs() -> anyhow::Result<()> {
     let sender_id = ACCOUNT_ID_SENDER
@@ -471,12 +417,12 @@ fn test_get_exactly_8_inputs() -> anyhow::Result<()> {
             begin
                 exec.prologue::prepare_transaction
 
-                # execute the `get_inputs` procedure to trigger note inputs number assertion
+                # execute the `get_inputs` procedure to trigger note inputs length assertion
                 push.0 exec.note::get_inputs
                 # => [num_inputs, 0]
 
-                # assert that the number if inputs is 8
-                push.8 assert_eq.err=\"number of inputs should be equal to 8\"
+                # assert that the inputs length is 8
+                push.8 assert_eq.err=\"number of inputs values should be equal to 8\"
 
                 # clean the stack
                 drop